+++ core/src/main/java/org/web3j/protocol/ObjectMapperFactory.java
@@ -74,6 +74,7 @@ public class ObjectMapperFactory {
 
         objectMapper.configure(JsonParser.Feature.ALLOW_UNQUOTED_FIELD_NAMES, true);
         objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        objectMapper.configure(DeserializationFeature.USE_BIG_DECIMAL_FOR_FLOATS, true);
 
         return objectMapper;
     }